+++ src/Contracts/ChainsNestedItemsInterface.php
@@ -8,6 +8,13 @@ namespace Michaels\Manager\Contracts;
  */
 interface ChainsNestedItemsInterface
 {
+
+    /**
+     * Deletes item at the current level of nesting (and below)
+     * @return mixed
+     */
+    public function drop();
+
     /**
      * Sets the current level of nesting.
      *
@@ -38,4 +45,19 @@ interface ChainsNestedItemsInterface
      * @return mixed item value
      */
     public function __call($name, $arguments);
+
+    /**
+     * Sets an item at the current nest level.
+     *
+     * Similar to __call() above. Must navigate to desired level through methods
+     * <$manager->one()->two()...> and then you can set a value as a property
+     * <$manger->one()->two = "two">
+     *
+     * See src/Traits/ChainsNestedItemsTrait.php for implementation example.
+     * @see Michaels\Manager\Contracts\ChainsNestedItemsInterface
+     * @param string $key The alias to be retrieved
+     * @param mixed $value Value to be set
+     * @return $this
+     */
+    public function __set($key, $value);
 }