+++ src/Manager.php
@@ -9,6 +9,7 @@ use JsonSerializable;
 use Interop\Container\ContainerInterface;
 use Michaels\Manager\Contracts\ChainsNestedItemsInterface;
 use Michaels\Manager\Contracts\ManagesItemsInterface;
+use Michaels\Manager\Traits\ArrayableTrait;
 use Michaels\Manager\Traits\ChainsNestedItemsTrait;
 use Michaels\Manager\Traits\ManagesItemsTrait;
 
@@ -29,7 +30,7 @@ class Manager implements
     IteratorAggregate,
     JsonSerializable
 {
-    use ManagesItemsTrait, ChainsNestedItemsTrait;
+    use ManagesItemsTrait, ChainsNestedItemsTrait, ArrayableTrait;
 
     /**
      * The items stored in the manager
@@ -44,61 +45,5 @@ class Manager implements
     public function __construct($items = [])
     {
         $this->initManager($items);
-    }
-
-    /**
-     * @implements ArrayAccess
-     */
-    public function offsetExists($offset)
-    {
-        return $this->exists($offset);
-    }
-
-    /**
-     * @implements ArrayAccess
-     */
-    public function offsetGet($offset)
-    {
-        return $this->get($offset);
-    }
-
-    /**
-     * @implements ArrayAccess
-     */
-    public function offsetSet($offset, $value)
-    {
-        $this->set($offset, $value);
-    }
-
-    /**
-     * @implements ArrayAccess
-     */
-    public function offsetUnset($offset)
-    {
-        $this->remove($offset);
-    }
-
-    /**
-     * @implements IteratorAggregate
-     */
-    public function getIterator()
-    {
-        return new ArrayIterator($this->getAll());
-    }
-
-    /**
-     * @implements Countable
-     */
-    public function count()
-    {
-        return count($this->items);
-    }
-
-    /**
-     * @implements JSONSerializable
-     */
-    public function jsonSerialize()
-    {
-        return $this->items;
     }
 }