+++ tests/Stubs/ManagesItemsTraitStub.php
@@ -1,13 +1,14 @@
 <?php
 namespace Michaels\Manager\Test\Stubs;
 
+use Michaels\Manager\Contracts\ManagesItemsInterface;
 use Michaels\Manager\Traits\ManagesItemsTrait;
 
 /**
  * Class ManagesItemsTraitStub
  * @package Stubs
  */
-class ManagesItemsTraitStub
+class ManagesItemsTraitStub implements ManagesItemsInterface
 {
     use ManagesItemsTrait;